requirement.txt
Flask == 3.0.2
flask-smorest == 0.43.0
python-dotenv == 1.0.1
openai == 1.6.1